11/23 #4

+++ final-pjt-back/movies/views.py
@@ -38,17 +38,19 @@ def movie_detail(request, movie_pk):
 
 
 @api_view(['GET', 'DELETE', 'PUT'])
+@permission_classes([IsAuthenticated])
 def comment_detail(request, comment_pk):
     # comment = Comment.objects.get(pk=comment_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
-
+    user = request.user
     if request.method == 'GET':
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if user.pk == comment.user_id:    
+            comment.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
         serializer = CommentSerializer(comment, data=request.data)
